+++ src/main/kotlin/sakirabusinesslabs/tournamentstorage/repository/sport/SportRepository.kt
@@ -3,5 +3,5 @@ package sakirabusinesslabs.tournamentstorage.repository.sport
 import org.springframework.data.jpa.repository.JpaRepository
 import sakirabusinesslabs.tournamentstorage.repository.sport.entity.Sport
 
-interface SportRepository: JpaRepository<Sport, Long> {
+interface SportRepository: JpaRepository<Sport, Int> {
 }
