+++ src/main/kotlin/sakirabusinesslabs/tournamentstorage/repository/tournament/TournamentRepository.kt
@@ -3,5 +3,5 @@ package sakirabusinesslabs.tournamentstorage.repository.tournament
 import org.springframework.data.jpa.repository.JpaRepository
 import sakirabusinesslabs.tournamentstorage.repository.tournament.entity.Tournament
 
-interface TournamentRepository: JpaRepository<Tournament, Long> {
+interface TournamentRepository: JpaRepository<Tournament, Int> {
 }
