+++ src/main/kotlin/sakirabusinesslabs/tournamentstorage/repository/tournamentMaker/TournamentMakerRepository.kt
@@ -3,5 +3,5 @@ package sakirabusinesslabs.tournamentstorage.repository.tournamentMaker
 import org.springframework.data.jpa.repository.JpaRepository
 import sakirabusinesslabs.tournamentstorage.repository.tournamentMaker.entity.TournamentMaker
 
-interface TournamentMakerRepository: JpaRepository<TournamentMaker, Long> {
+interface TournamentMakerRepository: JpaRepository<TournamentMaker, Int> {
 }
