+++ src/main/kotlin/sakirabusinesslabs/tournamentstorage/repository/user/UserRepository.kt
@@ -3,4 +3,4 @@ package sakirabusinesslabs.tournamentstorage.repository.user
 import org.springframework.data.jpa.repository.JpaRepository
 import sakirabusinesslabs.tournamentstorage.repository.user.entity.User
 
-interface UserRepository: JpaRepository<User, Long>
+interface UserRepository: JpaRepository<User, Int>
